// app/modules/reporting/models/api/response/enums/report-result-metadata-type.ts
export enum ReportResultMetadataType {
  GENERIC = 'generic',
  TIME_ZONE = 'timeZone',
  SECURITY_ANALYSIS = 'securityAnalysis',
  INTERFACE_CAPACITY = 'interfaceCapacity',
  DASHBOARD = 'DashboardKey',
  SITE = 'site',
  NUMERIC = 'numeric',
  MOMENT = 'moment',
  INSIGHT_KEY = 'insightsKey',
  DEVICE = 'device',
}